+++ packages/sdk-router/src/rfq/fastBridgeRouterSet.ts
@@ -1,6 +1,7 @@
 import { Provider } from '@ethersproject/abstract-provider'
 import { BigNumber } from '@ethersproject/bignumber'
 import invariant from 'tiny-invariant'
+import { Zero } from '@ethersproject/constants'
 
 import {
   BigintIsh,
@@ -14,6 +15,7 @@ import {
   SynapseModule,
   SynapseModuleSet,
   createNoSwapQuery,
+  applySlippageToQuery,
 } from '../module'
 import { FastBridgeRouter } from './fastBridgeRouter'
 import { ChainProvider } from '../router'
@@ -34,6 +36,9 @@ export class FastBridgeRouterSet extends SynapseModuleSet {
   public providers: {
     [chainId: number]: Provider
   }
+
+  // The answer to life, the universe, and everything
+  private readonly GAS_REBATE_FLAG = '0x2a'
 
   constructor(chains: ChainProvider[]) {
     super()
@@ -63,6 +68,21 @@ export class FastBridgeRouterSet extends SynapseModuleSet {
     const medianTime = MEDIAN_TIME_RFQ[chainId as keyof typeof MEDIAN_TIME_RFQ]
     invariant(medianTime, `No estimated time for chain ${chainId}`)
     return medianTime
+  }
+
+  /**
+   * @inheritdoc SynapseModuleSet.getGasDropAmount
+   */
+  public async getGasDropAmount(bridgeRoute: BridgeRoute): Promise<BigNumber> {
+    // TODO: test this once chainGasAmount is set to be non-zero
+    if (
+      bridgeRoute.destQuery.rawParams
+        .toLowerCase()
+        .startsWith(this.GAS_REBATE_FLAG)
+    ) {
+      return this.getFastBridgeRouter(bridgeRoute.destChainId).chainGasAmount()
+    }
+    return Zero
   }
 
   /**
@@ -146,6 +166,46 @@ export class FastBridgeRouterSet extends SynapseModuleSet {
     return {
       originPeriod: TEN_MINUTES,
       destPeriod: ONE_HOUR,
+    }
+  }
+
+  /**
+   * @inheritdoc SynapseModuleSet.applySlippage
+   */
+  public applySlippage(
+    originQueryPrecise: Query,
+    destQueryPrecise: Query,
+    slipNumerator: number,
+    slipDenominator: number
+  ): { originQuery: Query; destQuery: Query } {
+    // Max slippage for origin swap is 5% of the fixed fee
+    // Relayer is using a 10% buffer for the fixed fee, so if origin swap slippage
+    // is under 5% of the fixed fee, the relayer will still honor the quote.
+    let maxOriginSlippage = originQueryPrecise.minAmountOut
+      .sub(destQueryPrecise.minAmountOut)
+      .div(20)
+    // TODO: figure out a better way to handle destAmount > originAmount
+    if (maxOriginSlippage.isNegative()) {
+      maxOriginSlippage = BigNumber.from(0)
+    }
+    const originQuery = applySlippageToQuery(
+      originQueryPrecise,
+      slipNumerator,
+      slipDenominator
+    )
+    if (
+      originQuery.minAmountOut
+        .add(maxOriginSlippage)
+        .lt(originQueryPrecise.minAmountOut)
+    ) {
+      originQuery.minAmountOut =
+        originQueryPrecise.minAmountOut.sub(maxOriginSlippage)
+    }
+    // Never modify the dest query, as the exact amount from it will always be used by the Relayer
+    // So applying slippage there will only reduce the user proceeds on the destination chain
+    return {
+      originQuery,
+      destQuery: destQueryPrecise,
     }
   }