+++ packages/sdk-router/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.18](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.17...@synapsecns/sdk-router@0.3.18) (2024-01-12)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
+## [0.3.17](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.16...@synapsecns/sdk-router@0.3.17) (2024-01-11)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
 ## [0.3.16](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.15...@synapsecns/sdk-router@0.3.16) (2024-01-11)
 
 **Note:** Version bump only for package @synapsecns/sdk-router